Refactor LayoutProvider
