+++ src/models/ChangePasswordPayload.ts
@@ -1,4 +1,4 @@
-import { PlatformInfo } from '@libs/platformjs';
+import { PlatformInfo } from '@utils/platform-utils';
 
 export type ChangePasswordPayload = PlatformInfo & {
   uid: string;
